Refactor: move FileCertificateStore imp and interface to Udap.Common\Certificates
